+++ trend-trading-backtest-service/src/main/java/com/stefanie/trend/pojo/IndexData.java
@@ -2,9 +2,11 @@ package com.stefanie.trend.pojo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class IndexData {
     String date;
-    float closePoint;
+    BigDecimal closePoint;
 
 }
