+++ trend-trading-backtest-service/src/main/java/com/stefanie/trend/pojo/Profit.java
@@ -2,10 +2,12 @@ package com.stefanie.trend.pojo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Profit {
  
     String date;
-    float value;
+    BigDecimal value;
 
 }
